display titles types finance

+++ app/src/main/java/com/example/businesshelper/addActivity/MinusExpenseActivity.kt
@@ -62,7 +62,7 @@ class MinusExpenseActivity : AppCompatActivity() {
                 if(v.isChecked){
                     val exp = Expenses(
                         date = sdf.format(Date()),
-                        sum = edSum.text.toString().toInt(),
+                        sum = edSum.text.toString().toLong(),
                         type = "te${i+1}"
                     )
                     list.add(exp)
